requirements.txt
djangorestframework==3.14.0
channels==4.1.0
daphne==4.1.2 
django-redis==5.4.0
Django==4.2.7
requests==2.26.0